chore: move ng pages to new 'pages' dir
