+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/contentHash/EditablePhotoEntry.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from "react";
 import logRender from "@/logRender";
 import type { PhotoDbEntry } from "dropbox-hacking-photo-manager-shared";
-import EditableTextField from "../../day/editableTextField";
+import EditableTextField from "@/day/editableTextField";
 
 const EditablePhotoEntry = ({
   contentHash,
