+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/day/MaybeVisibleThumbnail.tsx
@@ -1,6 +1,6 @@
-import { useThumbnailLoader } from "../../context/thumbnails";
+import { useThumbnailLoader } from "@/context/thumbnails";
 import logRender from "@/logRender";
-import { useThumbnail } from "../../context/thumbnails/useThumbnail";
+import { useThumbnail } from "@/context/thumbnails/useThumbnail";
 import React from "react";
 import type { NamedFile } from "dropbox-hacking-photo-manager-shared";
