+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/fileId/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 
 import logRender from "@/logRender";
-import Navigate from "../../days/navigate";
+import Navigate from "@/days/navigate";
 import type { FileIdResult } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
 import {
