+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/fsck/index.tsx
@@ -1,24 +1,24 @@
 import React, { useEffect } from "react";
 
 import logRender from "@/logRender";
-import Navigate from "../../days/navigate";
-import type { BasicCountsType } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
+import Navigate from "@/days/navigate";
+import type { FsckType } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
 
-const BasicCounts = () => {
-  const latestValue = useLatestValueFromServerFeed<BasicCountsType>({
-    type: "rx.ng.basic-counts",
+const Fsck = () => {
+  const latestValue = useLatestValueFromServerFeed<FsckType>({
+    type: "rx.ng.fsck",
   });
 
   useEffect(() => {
-    document.title = "DPMNG - Basic counts";
+    document.title = "DPMNG - fsck";
   }, []);
 
   return (
     <>
       <Navigate />
 
-      <h1>Basic Counts</h1>
+      <h1>fsck</h1>
 
       {latestValue ? (
         <pre>{JSON.stringify(latestValue ?? null, null, 2)}</pre>
@@ -29,4 +29,4 @@ const BasicCounts = () => {
   );
 };
 
-export default logRender(BasicCounts);
+export default logRender(Fsck);
