+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/list-of-days/without-samples.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 
 import logRender from "@/logRender";
-import Navigate from "../../days/navigate";
+import Navigate from "@/days/navigate";
 import type { NGDaysNoSamplesType } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
 import SamePageLink from "@/components/samePageLink";
